+++ CHANGELOG.markdown
@@ -2,7 +2,8 @@
 
 ## Master (Unreleased)
 
-* Run closest .irbrc config while initializing IRB adapters.
+* Run closest .irbrc config while initializing the IRB adapters.
+* Add config.web_console.prevent_irbrc_execution option.
 
 ## 0.1.0 (July 30, 2013)